requirements.txt
altgraph==0.17.3
autopep8==2.0.2
blinker==1.6.2
click==8.1.4
colorama==0.4.6
debugpy==1.6.7
et-xmlfile==1.1.0
Flask==2.3.2
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
openpyxl==3.1.2
pefile==2023.2.7
Pillow==9.5.0
pycodestyle==2.10.0
pyinstaller==5.13.0
pyinstaller-hooks-contrib==2023.3
PyPDF2==3.0.1
python-dotenv==1.0.0
pywin32-ctypes==0.2.1
reportlab==4.0.4
tomli==2.0.1
typing_extensions==4.6.3
urllib3==2.0.3
uuid==1.30
Werkzeug==2.3.6
zipp==3.15.0
